detection.py: letting image be None for Detection class.
We should allow the instantiation of the Detection class to be
independent of the image, so we can, in the first place, only load the
model in memory then changes the image attribute whenever we need to run
a new detection with the same model. We thus remove the hassle of
charging the model everytime.